refactor with scroll component
