--- robofriends/src/App.js
@@ -1,36 +1,0 @@
-import React from "react";
-import CardList from "./CardList";
-import SearchBox from "./SearchBox";
-import { robots } from "./robots";
-import "./App.css";
-
-class App extends React.Component {
-  constructor() {
-    super();
-    this.state = {
-      robots: robots,
-      seachField: "",
-    };
-  }
-
-  onSearchChange = (event) => {
-    this.setState({ seachField: event.target.value });
-  };
-
-  render() {
-    const filteredRobots = this.state.robots.filter((robots) => {
-      return robots.name
-        .toLowerCase()
-        .includes(this.state.seachField.toLowerCase());
-    });
-    return (
-      <div className="tc">
-        <h1 className="f1">Robo Friends Cards</h1>
-        <SearchBox searchChange={this.onSearchChange}></SearchBox>
-        <CardList robots={filteredRobots}></CardList>
-      </div>
-    );
-  }
-}
-
-export default App;
