+++ robofriends/src/components/Scroll.js
@@ -1,0 +1,17 @@
+import React from "react";
+
+const Scroll = (props) => {
+  return (
+    <div
+      style={{
+        overflowY: "scroll",
+        border: "5px solid black",
+        height: "800px",
+      }}
+    >
+      {props.children}
+    </div>
+  );
+};
+
+export default Scroll;
